backend/functions/create-ticket/requirements.txt
functions-framework==3.4.0
google-cloud-firestore==2.11.1
google-cloud-language==2.9.1